+++ src/Exception/DriverException.php
@@ -6,7 +6,7 @@ namespace JSONAPI\Exception;
 
 class DriverException extends JsonApiException
 {
-    const DRIVER_UNKNOWN = 20;
-    const DRIVER_ANNOTATION_NOT_ON_GETTER = 21;
+    const UNKNOWN = 20;
+    const ANNOTATION_NOT_ON_GETTER = 21;
 
 }