+++ src/Exception/EncoderException.php
@@ -6,7 +6,7 @@ namespace JSONAPI\Exception;
 
 class EncoderException extends JsonApiException
 {
-    const ENCODER_UNKNOWN = 40;
-    const ENCODER_INVALID_FIELD = 43;
-    const ENCODER_CLASS_NOT_EXIST = 44;
+    const UNKNOWN = 40;
+    const INVALID_FIELD = 43;
+    const CLASS_NOT_EXIST = 44;
 }