+++ src/Exception/FactoryException.php
@@ -6,7 +6,7 @@ namespace JSONAPI\Exception;
 
 class FactoryException extends JsonApiException
 {
-    const FACTORY_UNKNOWN = 10;
-    const FACTORY_CLASS_IS_NOT_RESOURCE = 11;
-    const FACTORY_PATH_IS_NOT_VALID = 12;
+    const UNKNOWN = 10;
+    const CLASS_IS_NOT_RESOURCE = 11;
+    const PATH_IS_NOT_VALID = 12;
 }